created the order bean and implemented create order along with cart emptying function

+++ E-Commerce/src/main/java/com/masai/beans/OrderDTO.java
@@ -18,7 +18,7 @@ public class OrderDTO {
 	
 	private double gst;
 	
-	private int deliveryCost;
+	private double deliveryCost;
 	
 	private double totalCost;
